feat: convert modules to plugin
