+++ android/android-annotation/build.gradle
@@ -1,3 +1,5 @@
+import br.com.zup.beagle.Dependencies
+
 /*
  * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
  *
@@ -20,9 +22,9 @@ sourceCompatibility = JavaVersion.VERSION_1_8
 targetCompatibility = JavaVersion.VERSION_1_8
 
 dependencies {
-    implementation GeneralLibraries.kotlin
+    implementation Dependencies.GeneralLibraries.kotlin
 
-    implementation MoshiLibraries.moshi
+    implementation Dependencies.MoshiLibraries.moshi
 }
 
 //apply from: rootProject.file('maven-publish.gradle')
