--- android/buildSrc/.gitignore
@@ -1,1 +1,0 @@
-/build
