+++ android/gradle.properties
@@ -23,7 +23,6 @@ android.enableJetifier=true
 org.gradle.parallel=true
 org.gradle.daemon=true
 org.gradle.caching= true
-android.enableSeparateAnnotationProcessing=true
 
 
 # Publication
