+++ detekt.gradle
@@ -17,5 +17,5 @@
 apply plugin: 'io.gitlab.arturbosch.detekt'
 
 detekt {
-    config = files("$rootDir/config/detekt-default-config.yml")
+    config = files("$rootDir/../config/detekt-default-config.yml")
 }
